simplification and cleanup
